Back Pressing Issues Fixed
Pressing back on any activity’s UI (except for HomeActivity), takes the
user back to the HomeActivity. When back pressed from the HomeActivity,
the app is minimized.
Also, instead of keeping the activities’ launchMode as
‘singleInstance’, a better approach is implemented to set the intent
flag to bring the existing instance on the top of the stack and thus
display it.
We need the HomeActivity’s launchMode to be ‘singleTask’ to enable
proper press-back case.